requirements.txt
pyjwt
fastapi
passlib[bcrypt]
pydantic